Update array dimension type from `uint32_t` to `uint64_t` for improved range support in type management.

+++ src/Sema/Type/TypeInfo.cpp
@@ -296,7 +296,7 @@ TypeInfo TypeInfo::makeSlice(TypeRef pointeeTypeRef, TypeInfoFlags flags)
     return ti;
 }
 
-TypeInfo TypeInfo::makeArray(const std::vector<uint32_t> &dims, TypeRef elementTypeRef, TypeInfoFlags flags)
+TypeInfo TypeInfo::makeArray(const std::vector<uint64_t> &dims, TypeRef elementTypeRef, TypeInfoFlags flags)
 {
     TypeInfo ti{TypeInfoKind::Array, flags};
     std::construct_at(&ti.asArray.dims, dims);
